originLambda/requirements.txt
Pillow~=7.1.2
python-resize-image~=1.1.19
pymsteams~=0.1.13
boto3~=1.13.14
botocore~=1.16.14